Slight re-org
The client had enough extra stuff in it that it made sense to start
separating things into different files
