+++ nmagent/client.go
@@ -18,42 +18,6 @@ const (
 	GetNetworkConfigPath string = "/machine/plugins/?comp=nmagent&type=NetworkManagement/joinedVirtualNetworks/%s/api-version/1"
 )
 
-// Error is a aberrent condition encountered when interacting with the NMAgent
-// API
-type Error struct {
-	Code int // the HTTP status code received
-}
-
-func (e Error) Error() string {
-	return fmt.Sprintf("nmagent: http status %d", e.Code)
-}
-
-// Temporary reports whether the error encountered from NMAgent should be
-// considered temporary, and thus retriable
-func (e Error) Temporary() bool {
-	return e.Code == http.StatusProcessing
-}
-
-type VirtualNetwork struct {
-	CNetSpace      string   `json:"cnetSpace"`
-	DefaultGateway string   `json:"defaultGateway"`
-	DNSServers     []string `json:"dnsServers"`
-	Subnets        []Subnet `json:"subnets"`
-	VNetSpace      string   `json:"vnetSpace"`
-	VNetVersion    string   `json:"vnetVersion"`
-}
-
-type Subnet struct {
-	AddressPrefix string `json:"addressPrefix"`
-	SubnetName    string `json:"subnetName"`
-	Tags          []Tag  `json:"tags"`
-}
-
-type Tag struct {
-	Name string `json:"name"`
-	Type string `json:"type"` // the type of the tag (e.g. "System" or "Custom")
-}
-
 // Client is an agent for exchanging information with NMAgent
 type Client struct {
 	HTTPClient *http.Client
