+++ nmagent/responses.go
@@ -1,0 +1,21 @@
+package nmagent
+
+type VirtualNetwork struct {
+	CNetSpace      string   `json:"cnetSpace"`
+	DefaultGateway string   `json:"defaultGateway"`
+	DNSServers     []string `json:"dnsServers"`
+	Subnets        []Subnet `json:"subnets"`
+	VNetSpace      string   `json:"vnetSpace"`
+	VNetVersion    string   `json:"vnetVersion"`
+}
+
+type Subnet struct {
+	AddressPrefix string `json:"addressPrefix"`
+	SubnetName    string `json:"subnetName"`
+	Tags          []Tag  `json:"tags"`
+}
+
+type Tag struct {
+	Name string `json:"name"`
+	Type string `json:"type"` // the type of the tag (e.g. "System" or "Custom")
+}
